Convert apply/concat tests to mocha
